% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{chat_extract_hunter_email_address}
\alias{chat_extract_hunter_email_address}
\title{Extract Hunter Email Address}
\usage{
chat_extract_hunter_email_address(
  chat,
  company_domain,
  company_name,
  first_name,
  last_name
)
}
\arguments{
\item{chat}{An \code{ellmer::chat} object.}

\item{company_domain}{A character string representing the company domain.}

\item{company_name}{A character string representing the company name.}

\item{first_name}{A character string representing the first name of the person.}

\item{last_name}{A character string representing the last name of the person.}
}
\value{
A list containing the email address information.
}
\description{
Extract the email address from the chat response using the Hunter.io API.
}
