% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_address}
\alias{clean_address}
\title{Clean Address}
\usage{
clean_address(address)
}
\arguments{
\item{address}{A character string representing the street address.}
}
\value{
A character string representing the cleansed and formatted address.
}
\description{
Cleanse and format an address in a standardized manner by removing unnecessary
elements and characters and combining the address, city, state, and zip code
components into a single string.
}
