% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_company_name}
\alias{clean_company_name}
\title{Clean Company Name}
\usage{
clean_company_name(company_name)
}
\arguments{
\item{company_name}{Character string representing the company name to be cleaned.}
}
\value{
A character string representing the cleansed and formatted company name.
}
\description{
Cleanse and format a company name by removing common suffixes and
non-alphanumeric characters, and converting it to lowercase.
}
