% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_first_last_name}
\alias{clean_first_last_name}
\title{Clean First and Last Name}
\usage{
clean_first_last_name(first_name, last_name)
}
\arguments{
\item{first_name}{Character string representing the first name to be cleaned.}

\item{last_name}{Character string representing the last name to be cleaned.}
}
\value{
A character string representing the cleansed and formatted full name.
}
\description{
Cleanse and format a first and last name by removing any trailing initials.
}
