% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{discover_owner_emails}
\alias{discover_owner_emails}
\title{Discover Owner Emails}
\usage{
discover_owner_emails(owners)
}
\arguments{
\item{owners}{A data frame containing owner information.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{discovered_domain}: The domain discovered from Google Maps.
\item \code{discovered_email}: The email address discovered or generated.
\item \code{match_quality}: The quality of the match (if applicable).
\item \code{email_score}: The score of the discovered email (if applicable).
\item \code{business_phone}: The business phone number.
\item \code{business_type}: The type of business.
\item \code{business_status}: The status of the business.
\item \code{business_website}: The business website.
\item \code{business_maps_url}: The Google Maps URL for the business.
\item \code{business_latitude}: The latitude of the business location.
\item \code{business_longitude}: The longitude of the business location.
\item \code{processing_status}: The status of the processing for each record.
\item \code{final_email}: The final email address, either existing or discovered/generated.
}
}
\description{
This function attempts to discover email addresses for a list of owners by
searching for their business information using Google Maps and then using
Hunter.io to find email addresses based on the discovered domain and
owner first and last name.
}
\details{
This function processes a data frame of owners, each with their company name,
address, phone number, and optionally an existing email address. It first
searches for the business information using Google Maps, retrieves the domain,
and then attempts to find an email address using Hunter.io. If no email is found,
it generates a potential email address based on the owner's name and the discovered domain.
}
\examples{
\dontrun{
hud_owners_data <- data("hud_owners")
hud_owners_sample <- dplyr::slice_sample(hud_owners_data, n = 5)
discovered_emails <- discover_owner_emails(hud_owners_sample)
}
}
