% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{email_validation}
\alias{email_validation}
\alias{validate_email_regex}
\alias{validate_email_mx}
\alias{validate_domain}
\title{Email Address and Domain Server Validation}
\usage{
validate_email_regex(email)

validate_email_mx(email, timeout = 5)

validate_domain(domain)
}
\arguments{
\item{email}{A character string representing the email address to validate.}

\item{timeout}{An integer representing the timeout in seconds for DNS queries.}

\item{domain}{A character string representing the domain to validate.}
}
\value{
\itemize{
\item \code{validate_email_regex()}: Returns \code{TRUE} if the email address is valid according to the regex pattern, otherwise \code{FALSE}.
\item \code{validate_email_mx()}: Returns \code{TRUE} if the email address has valid MX records or A records, otherwise \code{FALSE}.
\item \code{validate_domain()}: Returns \code{TRUE} if the domain has valid A and MX records, otherwise \code{FALSE}.
}
}
\description{
Validate email addresses and domain servers by checking
format and regular expression patterns, \code{MX} DNS records, and \code{A} DNS
records.

Functions:
\itemize{
\item \code{validate_email_regex()}: Validates the format of an email address using a regular expression.
\item \code{validate_email_mx()}: Validates the email address by checking its MX records and falling back to A records if necessary.
\item \code{validate_domain()}: Validates the domain by checking its A and MX records.
}
}
\examples{
email <- "jimmy.briggs@noclocks.dev"
validate_email_regex(email)
validate_email_mx(email)
domain <- sub(".*@", "", email)
validate_domain(domain)
}
