% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_domain_from_url}
\alias{get_domain_from_url}
\title{Get Domain from URL}
\usage{
get_domain_from_url(url)
}
\arguments{
\item{url}{A character string representing the URL to extract the domain from.}
}
\value{
A character string representing the domain extracted from the URL.
}
\description{
Extract the domain from a URL string by removing the protocol and any sub-domains.
}
