% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_full_address}
\alias{get_full_address}
\title{Get Full Address}
\usage{
get_full_address(address, city, state, zip)
}
\arguments{
\item{address}{A character string representing the street address.}

\item{city}{A character string representing the city name.}

\item{state}{A character string representing the state abbreviation.}

\item{zip}{A character string representing the zip code.}
}
\value{
A character string representing the full address.
}
\description{
Combine the address, city, state, and zip code components into a single string.
}
