% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmaps.R
\name{gmaps_get_place_details}
\alias{gmaps_get_place_details}
\title{Get Place Details}
\usage{
gmaps_get_place_details(place_id, api_key = get_gmaps_api_key())
}
\arguments{
\item{place_id}{A character string representing the place ID.}

\item{api_key}{A character string representing the Google Maps API key.
Will default to the result from \code{\link[=get_gmaps_api_key]{get_gmaps_api_key()}}.}
}
\value{
A list containing the following elements:
\itemize{
\item \code{website}: A character string representing the website URL.
\item \code{address}: A character string representing the formatted address.
\item \code{phone}: A character string representing the formatted phone number.
\item \code{name}: A character string representing the name of the place.
}
}
\description{
Get the details of a place using the Google Maps Places API.
}
