% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmaps.R
\name{gmaps_get_place_id}
\alias{gmaps_get_place_id}
\title{Get Google Maps Place ID}
\usage{
gmaps_get_place_id(company_name, lat = NULL, lon = NULL)
}
\arguments{
\item{company_name}{Character string representing the name of the company.}

\item{lat}{Latitude of the location.}

\item{lon}{Longitude of the location.}
}
\value{
The place ID of the company.
}
\description{
Get the place ID for a given address using the Google Maps Places API.
}
