% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hunterio.R
\name{hunter_get_email_address}
\alias{hunter_get_email_address}
\title{Get Email Address via Hunter.io}
\usage{
hunter_get_email_address(
  company_domain,
  company_name,
  first_name,
  last_name,
  api_key = get_hunterio_api_key()
)
}
\arguments{
\item{company_domain}{Character string representing the company domain.}

\item{company_name}{Character string representing the company name.}

\item{first_name}{Character string representing the person's first name.}

\item{last_name}{Character string representing the person's last name.}

\item{api_key}{Hunter.io API key. If not provided, uses \code{\link[=get_hunterio_api_key]{get_hunterio_api_key()}} to retrieve it.}
}
\value{
A list containing the following elements:
\itemize{
\item \code{first_name}: The first name of the person.
\item \code{last_name}: The last name of the person.
\item \code{email}: The retrieved email address.
\item \code{score}: The score of the email address.
\item \code{domain}: The company domain.
\item \code{company}: The company name.
\item \code{phone_number}: The phone number associated with the email address.
\item \code{verification_date}: The date of email verification.
}
}
\description{
This function retrieves an email address based on the provided company domain, company name,
and person's first and last name using the Hunter.io API.
}
