% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hunterio.R
\name{hunter_parse_email_response}
\alias{hunter_parse_email_response}
\title{Parse Hunter.io Email Response}
\usage{
hunter_parse_email_response(resp_json)
}
\arguments{
\item{resp_json}{The JSON response from the Hunter.io API.}
}
\value{
A list containing the following elements:
\itemize{
\item \code{first_name}: The first name of the person.
\item \code{last_name}: The last name of the person.
\item \code{email}: The retrieved email address.
\item \code{score}: The score of the email address.
\item \code{domain}: The company domain.
\item \code{company}: The company name.
\item \code{phone_number}: The phone number associated with the email address.
\item \code{verification_date}: The date of email verification.
}
}
\description{
This function parses the JSON response from the Hunter.io API for email discovery and verification.
}
