% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{string_similarity}
\alias{string_similarity}
\title{Calculate String Similarity}
\usage{
string_similarity(string1, string2)
}
\arguments{
\item{string1}{A character string representing the first string.}

\item{string2}{A character string representing the second string.}
}
\value{
A numeric value between 0 and 1 representing the similarity between the two strings.
}
\description{
Calculate the similarity between two strings using the Jaro-Winkler distance method.
}
