% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{tools}
\alias{tools}
\alias{tool_gmaps_places_search}
\alias{tool_hunter_get_email_address}
\title{Tools}
\usage{
tool_gmaps_places_search()

tool_hunter_get_email_address()
}
\value{
Each function returns an \code{\link[ellmer:tool]{ellmer::tool()}} definition object.
}
\description{
Register various tools for geocoding, address cleaning, and domain extraction.

Functions:
\itemize{
\item \code{tool_gmaps_places_search()}: Search for places using Google Maps Places API to find website and
contact information.
\item \code{tool_hunter_get_email_address()}: Retrieve an email address based on the provided company domain, company name,
and person's first and last name using the Hunter.io API.
}
}
