% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{types}
\alias{types}
\alias{type_owner_contact}
\alias{type_property_analysis}
\alias{type_email_draft}
\title{Types}
\usage{
type_owner_contact()

type_property_analysis()

type_email_draft()
}
\value{
Each function returns an \code{\link[ellmer:type_boolean]{ellmer::type_object()}} definition object.
}
\description{
Custom type definitions for the data structures used in the tools.

Functions:
\itemize{
\item \code{type_gmaps_geocode_address()}: Defines the structure for geocoded address data.
\item \code{type_gmaps_place_details()}: Defines the structure for place details data.
\item \code{type_clean_address()}: Defines the structure for cleaned address data.
\item \code{type_domain_from_url()}: Defines the structure for domain extracted from URL data.
\item \code{type_gmaps_results()}: Defines the structure for Google Maps results data.
}
}
