% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cspr-package.R
\docType{package}
\name{cspr-package}
\alias{cspr}
\alias{cspr-package}
\title{cspr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

}
\keyword{internal}
