% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{tools}
\alias{tools}
\alias{tool_clean_address}
\alias{tool_gmaps_geocode_address}
\alias{tool_gmaps_get_place_id}
\alias{tool_gmaps_get_place_details}
\alias{tool_get_domain_from_url}
\title{Tools}
\usage{
tool_clean_address()

tool_gmaps_geocode_address()

tool_gmaps_get_place_id()

tool_gmaps_get_place_details()

tool_get_domain_from_url()
}
\value{
Each function returns an \code{\link[ellmer:tool]{ellmer::tool()}} definition object.
}
\description{
Register various tools for geocoding, address cleaning, and domain extraction.

Functions:
\itemize{
\item \code{tool_clean_address()}: Cleans and formats an address.
\item \code{tool_gmaps_geocode_address()}: Geocodes a given address using Google Maps API.
\item \code{tool_gmaps_get_place_id()}: Gets the place ID for a given company name.
\item \code{tool_gmaps_get_place_details()}: Fetches detailed information about a place.
\item \code{tool_get_domain_from_url()}: Extracts the domain from a URL.
}
}
